fix: support SSO session credentials from aws login by checking for sessionToken before calling GetSessionToken
